LU-1187 lod: reorganize lod_ost
Move lod_ost to lod_tgt_desc, so LOD can be used
to manage OSP for both MDT and OST.

Reviewed-on: http://review.whamcloud.com/4921
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>

+++ lustre/include/lu_object.h
@@ -490,10 +490,10 @@ struct lu_object {
          * Depth. Top level layer depth is 0.
          */
         int                                lo_depth;
-        /**
-         * Flags from enum lu_object_flags.
-         */
-        unsigned long                      lo_flags;
+	/**
+	 * Flags from enum lu_object_flags.
+	 */
+	__u32					lo_flags;
         /**
          * Link to the device, for debugging.
          */
